Fixed the compiled object's cell's map/address initialization issue: before resetting or setting any values in the compiled object, all reward components cells have to be evaluated first in order to get initialized, and after setting the action of deployment increments, the reward component cells have to be re-evaluated again to take the updated/reset value into consideration when doing the formula calculation.
Now the env.py's output rewards components are (almost) exactly the same as Excel's workbook's total capex opex revenue emissions: In 'Pathways to Net Zero - Simplified.xlsx' - 'Gale' spreadsheet, simply apply an increments of [1.0 1.0 2.0] to offshore wind, blue hydrogen, and green hydrogen on top of 2030 values and set the new values as 2031 values, then apply constant increments [2.0 2.0 5.0] (as in the test_reference_environment_direct_deployment.py) for all following years, then in the 'Outputs' and 'CCUS', the total capex opex revenue and emissions from 2031 to 2050 are exactly the same as env.py's outputs. Moreover, even in the original 'Pathways to Net Zero.xlsx', when doing the same operation, the total opex revenue and emissions from 2031 to 2050 are still exactly the same as env.py's outputs, whereas the total capex is different because in the original 'Pathways to Net Zero.xlsx', offshore wind's capex depends on some numbers that are weighted sum of future values, which is not removed and replaced by current year's values only in the 'Pathways to Net Zero - Simplified.xlsx'.